revert to callbacks
