+++ src/index.ts
@@ -1,2 +1,4 @@
 export { generateOnRampURL } from './onramp/generateOnRampURL';
 export { initOnRamp } from './onramp/initOnRamp';
+export type { InitOnRampParams } from './onramp/initOnRamp';
+export type { CBPayInstanceType } from './utils/CBPayInstance';
