+++ tsup.config.ts
@@ -1,12 +1,7 @@
 import { defineConfig } from 'tsup';
 
 export default defineConfig({
-  entry: [
-    'src/index.ts',
-    'src/onramp/es5/initOnRamp.ts',
-    'src/utils/postMessage.ts',
-    'src/utils/CBPayInstance.ts',
-  ],
+  entry: ['src/index.ts', 'src/utils/postMessage.ts', 'src/utils/CBPayInstance.ts'],
   splitting: true,
   sourcemap: true,
   clean: true,
